+++ packages/uni-app/uni_modules/mxuni-router-v2/js_sdk/index.js
@@ -1,4 +1,4 @@
-export { AnimationPlugin, ChannelPlugin, InterceptorPlugin, NavigationFailure, ParamsPlugin, ROUTER_SYMBOL, RouterError, RouterErrorCode, UniApiError, UniEventChannel, createRouter, noopChannel, usePageChannel, useRoute, useRouter } from './chunk-6D5DCTIL.js';
+export { AnimationPlugin, ChannelPlugin, InterceptorPlugin, NavigationFailure, ParamsPlugin, ROUTER_SYMBOL, RouterError, RouterErrorCode, UniApiError, UniEventChannel, createRouter, noopChannel, onBeforeRouteLeave, usePageChannel, useRoute, useRouter } from './chunk-KMISK4GA.js';
 
 // src/types/route.ts
 var DEFAULT_ANIMATION_DURATION = 300;